Generar Cartas en formato pdf

+++ src/main/java/es/uniovi/asw/persistencia/Generador.java
@@ -1,6 +1,15 @@
 package es.uniovi.asw.persistencia;
 
 import java.io.*;
+
+import com.itextpdf.text.BaseColor;
+import com.itextpdf.text.Document;
+import com.itextpdf.text.DocumentException;
+import com.itextpdf.text.Font;
+import com.itextpdf.text.FontFactory;
+import com.itextpdf.text.Paragraph;
+import com.itextpdf.text.pdf.PdfWriter;
+
 import es.uniovi.asw.logica.Votante;
 
 /**
@@ -8,32 +17,74 @@ import es.uniovi.asw.logica.Votante;
  * 
  * @author Adrian
  */
-public class Generador {
+public class Generador implements GeneradorCartas {
+	
+	private String rutaCartasTxt = "cartasTXT/";
+	private String rutaCartasPdf = "cartasPDF/";
 		
+	@Override
+	public void generador(Votante votante) throws IOException, DocumentException {
+
+		generarCarta(votante);
+	}	
+	
 	/**
 	 * Constructor de la clase
 	 * @param votante
 	 * @throws IOException
+	 * @throws DocumentException 
 	 */
-	public Generador(Votante votante) throws IOException {
+	public Generador(Votante votante) throws IOException, DocumentException {
 		
 		generarCarta(votante);
+	}
+
+	private void generarCartaPdf(Votante votante, String pass) throws FileNotFoundException, DocumentException {
+		
+		String rutaCarta = rutaCartasPdf + votante.getNif() + ".pdf";
+
+		Document documento = new Document();
+		FileOutputStream ficheroPdf = new FileOutputStream(rutaCarta);
+					 
+		PdfWriter.getInstance(documento,ficheroPdf).setInitialLeading(20);
+		 
+		documento.open();
+		documento.add(new Paragraph("Usuario: " + votante.getNombre() + "." + votante.getEmail(),
+		                FontFactory.getFont("arial", 20, Font.BOLD)));       
+		
+		documento.add(new Paragraph(pass,
+                FontFactory.getFont("arial", 20, Font.BOLD)));
+		
+		documento.close();	
 	}
 
 	/**
 	 * Metodo principal que genera la carta
 	 * @param votante
 	 * @throws IOException 
+	 * @throws DocumentException 
 	 */
-	public void generarCarta(Votante votante) throws IOException {
+	public void generarCarta(Votante votante) throws IOException, DocumentException {
 
-        BufferedWriter carta = new BufferedWriter(new FileWriter(new 
-        		File(obtenerNombreFichero(votante) + ".txt")));
-            
-        carta.write("Usuario: " + generarUsuario(votante));
-		carta.write("\nClave: " + generarPassword(votante));
-        
-		carta.close();		
+		String rutaCarta = rutaCartasTxt + votante.getNif() + ".txt";
+
+		File fichero = new File(rutaCarta);
+		
+		if (!fichero.exists())
+		{
+	        BufferedWriter carta = new BufferedWriter(new FileWriter(new 
+	        		File(rutaCartasTxt + votante.getNif() + ".txt")));
+	            
+	        carta.write("Usuario: " + votante.getNombre() + "." + votante.getEmail());
+			String pass = "\nClave: " + generarPassword(votante);
+	        carta.write(pass);
+	        
+			carta.close();
+			
+			generarCartaPdf(votante, pass);
+		}
+		else
+			System.out.println("La carta para este usuario ya existe");
 	}
 	
 	/**
@@ -43,22 +94,8 @@ public class Generador {
 	 */
 	private String generarPassword(Votante votante) {
 
-		String pass = votante.getNombre() + "." + votante.getNif();
-		
-		pass += generarValoresClave();
-		
-		return pass;
-	}
-
-	/**
-	 * Método que genera un código aleatorio para aumentar la complejidad 
-	 * de la clave
-	 * @return codigo
-	 */
-	private String generarValoresClave() {
-
 		int i, valor;
-		char[] codigo = new char[5];
+		char[] codigo = new char[10];
 		char[] elementos = {'0','1','2','3','4','5','6','7','8','9',
 						'a','b','c','d','e','f','g','h','i','j','k','l','m','n',
 						'o','p','q','r','s','t','u','v','w','x','y','z',
@@ -73,24 +110,4 @@ public class Generador {
 		
 		return new String(codigo);
 	}
-
-	/**
-	 * Devuelve el nombre de la carta, que será su dni
-	 * @param votante
-	 * @return dni del votante
-	 */
-	private String obtenerNombreFichero(Votante votante) {
-		
-		return votante.getNif();
-	}
-
-	/**
-	 * Genera el usuario que estará compuesto por el nombre y el email
-	 * @param votante
-	 * @return usuario
-	 */
-	private String generarUsuario(Votante votante) {
-		
-		return votante.getNombre() + "." + votante.getEmail();
-	}	
 }
